Remove indeterminate progress
